We need to update the options for the new supply teachers framework so that we move the education technology platforms option to the ‘What is your school looking for?’ page. This is what has been done in this commit.
I have updated the specs and feature tests to account for these changes.
