+++ config/locales/views/supply_teachers.en.yml
@@ -707,7 +707,8 @@ en:
             all_suppliers: A list of all agencies
             calculate_fta_to_perm_fee: To work out what fixed-term worker transfer fee I could be charged
             calculate_temp_to_perm_fee: To calculate the fee my school will be charged if I make an agency worker permanent
-            managed_service_provider: An agency to manage all my school's needs; a 'managed service provider'
+            education_technology_platform: Education technology platform service
+            master_vendor: An agency to manage all my school's needs; a 'managed service provider - Master Vendor'
             worker: An agency who can provide my school with an individual worker
           page_title: What is your school looking for?
           question: What is your school looking for?
