+++ features/accessibility/supply_teachers/rm6238/journey/master_vendor_providers_accessibility.feature
@@ -3,29 +3,14 @@ Feature: Supply Teachers - Managed service providers - Accessibility
 
   Background: Navigate to What type of managed service do you want?
     Given I sign in and navigate to the start page for the 'RM6238' framework in 'supply teachers'
-    And I select "An agency to manage all my school's needs; a 'managed service provider'"
-    And I click on 'Continue'
-    Then I am on the 'What type of managed service do you want?' page
-
-  Scenario: What type of managed service do you want? page
-    Then the page should be axe clean
-
-  Scenario: Is your contract likely to be worth more than £2.5 million? page
-    And I select 'Master vendor managed service'
+    And I select "An agency to manage all my school's needs; a 'managed service provider - Master Vendor'"
     And I click on 'Continue'
     Then I am on the 'Is your contract likely to be worth more than £2.5 million?' page
-    Then the page should be axe clean
 
-  Scenario: Education technology platform service providers page
-    And I select 'Education technology platform service'
-    And I click on 'Continue'
-    Then I am on the 'Education technology platform service providers' page
+  Scenario: Is your contract likely to be worth more than £2.5 million? page
     Then the page should be axe clean
 
   Scenario Outline: Master vendor managed service providers page
-    And I select 'Master vendor managed service'
-    And I click on 'Continue'
-    Then I am on the 'Is your contract likely to be worth more than £2.5 million?' page
     And I select '<option>'
     And I click on 'Continue'
     Then I am on the 'Master vendor managed service providers' page
