+++ features/services/supply_teachers/rm6238/jounrey/master_vendor/master_vendor_above_threshold.feature
@@ -3,10 +3,7 @@ Feature: Supply Teachers - Master vendors - Above threshold
 
   Scenario: Master vendors - Above threshold results
     Given I sign in and navigate to the start page for the 'RM6238' framework in 'supply teachers'
-    And I select "An agency to manage all my school's needs; a 'managed service provider'"
-    And I click on 'Continue'
-    Then I am on the 'What type of managed service do you want?' page
-    And I select 'Master vendor managed service'
+    And I select "An agency to manage all my school's needs; a 'managed service provider - Master Vendor'"
     And I click on 'Continue'
     Then I am on the 'Is your contract likely to be worth more than £2.5 million?' page
     And I select 'Yes'
